Refactor EffectiveMarginalityCalculator and MarginalityCalculator for improved clarity and functionality
- Introduced a new EffectiveMarginalityCalculator class to encapsulate effective marginality calculations, separating it from the existing MarginalityCalculator.
- Updated the MarginalityCalculator to remove redundant effective marginality logic, streamlining its focus on general marginality calculations.
- Enhanced the EffectiveMarginalityCalculator with methods for calculating effective marginality levels and indicators based on defined thresholds.
- Adjusted imports in WeeklyFinancialReportCalculations to reference the new EffectiveMarginalityCalculator, ensuring consistent usage across the application.

These changes improve code organization and maintainability while enhancing the clarity of financial calculations.
