+++ workers/main/src/services/WeeklyFinancialReport/MarginalityCalculator.ts
@@ -1,8 +1,5 @@
 import {
-  HIGH_EFFECTIVE_MARGINALITY_THRESHOLD,
   HIGH_MARGINALITY_THRESHOLD,
-  LOW_EFFECTIVE_MARGINALITY_THRESHOLD,
-  MEDIUM_EFFECTIVE_MARGINALITY_THRESHOLD,
   MEDIUM_MARGINALITY_THRESHOLD,
 } from '../../configs/weeklyFinancialReport';
 
@@ -12,25 +9,11 @@ export enum MarginalityLevel {
   Low = 'low',
 }
 
-export enum EffectiveMarginalityLevel {
-  High = 'high',
-  Medium = 'medium',
-  Low = 'low',
-  VeryLow = 'veryLow',
-}
-
 export interface MarginalityResult {
   marginAmount: number;
   marginalityPercent: number;
   effectiveMarginalityIndicator: string;
   level: MarginalityLevel;
-}
-
-export interface EffectiveMarginalityResult {
-  marginAmount: number;
-  marginalityPercent: number;
-  effectiveMarginalityIndicator: string;
-  level: EffectiveMarginalityLevel;
 }
 
 export class MarginalityCalculator {
@@ -64,47 +47,6 @@ export class MarginalityCalculator {
       case MarginalityLevel.Low:
       default:
         return ':arrowdown:';
-    }
-  }
-}
-
-export class EffectiveMarginalityCalculator {
-  static calculate(revenue: number, cogs: number): EffectiveMarginalityResult {
-    const marginAmount = revenue - cogs;
-    const marginalityPercent = revenue > 0 ? (marginAmount / revenue) * 100 : 0;
-    const level = this.classify(marginalityPercent);
-    const effectiveMarginalityIndicator = this.getIndicator(level);
-
-    return {
-      marginAmount,
-      marginalityPercent,
-      effectiveMarginalityIndicator,
-      level,
-    };
-  }
-
-  static classify(percent: number): EffectiveMarginalityLevel {
-    if (percent >= HIGH_EFFECTIVE_MARGINALITY_THRESHOLD)
-      return EffectiveMarginalityLevel.High;
-    if (percent >= MEDIUM_EFFECTIVE_MARGINALITY_THRESHOLD)
-      return EffectiveMarginalityLevel.Medium;
-    if (percent >= LOW_EFFECTIVE_MARGINALITY_THRESHOLD)
-      return EffectiveMarginalityLevel.Low;
-
-    return EffectiveMarginalityLevel.VeryLow;
-  }
-
-  static getIndicator(level: EffectiveMarginalityLevel): string {
-    switch (level) {
-      case EffectiveMarginalityLevel.High:
-        return `:large_green_circle:`;
-      case EffectiveMarginalityLevel.Medium:
-        return `:large_yellow_circle:`;
-      case EffectiveMarginalityLevel.Low:
-        return `:red_circle:`;
-      case EffectiveMarginalityLevel.VeryLow:
-      default:
-        return `:no_entry:`;
     }
   }
 }
